Add support authenticated only useful link
Some links shown in the useful links section require authentication and therefore should not be shown to the public.

We decide which links show and not show based on their domain. This list is stored at the top of the [consideration.html](/planning-data-design/application/templates/consideration.html) file.

Please note: you should include the host, for example, `https://www.example.com/` would be `www.example.com`.

https://github.com/digital-land/data-design/issues/150